third_party/WebKit/LayoutTests/accessibility/contenteditable-caret-position-expected.txt

Line 1

Line 1
Line 2

Line 3

This tests that caret position is reported correctly for contenteditable elements.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS textboxAccessible.selectionStart is 0
PASS textboxAccessible.selectionEnd is 0
PASS textboxAccessible.selectionStart is 1
PASS textboxAccessible.selectionEnd is 1
PASS textboxAccessible.selectionStart is 0
PASS textboxAccessible.selectionEnd is 0
PASS line2Accessible.selectionStart is 0
PASS line2Accessible.selectionEnd is 0
PASS line2Accessible.selectionStart is 3
PASS line2Accessible.selectionEnd is 3
PASS contenteditableAccessible.selectionStart is 6
PASS contenteditableAccessible.selectionEnd is 6
PASS contenteditableAccessible.selectionStart is 13
PASS contenteditableAccessible.selectionEnd is 13
PASS contenteditableAccessible.selectionStart is 21
PASS contenteditableAccessible.selectionEnd is 21
PASS successfullyParsed is true

TEST COMPLETE
